Move route components to Containers and begin scaffolding main layout in App
