+++ src/Components/index.js
@@ -1,4 +1,1 @@
 export {default as Header} from "./Header"
-export {default as Projects} from "./Projects"
-export {default as ContactMe} from "./ContactMe"
-export {default as Blog} from "./Blog"
