+++ src/Containers/index.js
@@ -1,0 +1,3 @@
+export {default as Projects} from "./Projects"
+export {default as ContactMe} from "./ContactMe"
+export {default as Blog} from "./Blog"
